requirements.txt
google-cloud-aiplatform==1.35.0
watchdog==3.0.0
streamlit==1.28.0
langchain==0.0.325
google-cloud-aiplatform[preview]
google-cloud-secret-manager==2.16.4
vertexai==0.0.1